server/yolo_music/requirements.txt
Django>=4.2.0
django-cors-headers>=4.1.0
mutagen>=1.46.0
djangorestframework>=3.15.2